Also update tests/CMakeLists.txt if it exists
