+++ src/pf/fs.hpp
@@ -3,5 +3,6 @@
 
 #include <pf/fs/ascending_iterator.hpp>
 #include <pf/fs/core.hpp>
+#include <pf/fs/glob.hpp>
 
 #endif  // PF_FS_HPP_INCLUDED
